Updated Amiberry to v5.2. Amiberry setup script now checks if Amiberry ZIP-package contains subfolders.

+++ scripts/setup/setup_amiberry.sh
@@ -11,28 +11,28 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # Global variables to define the Amiberry binary file to download.
 GH_URL="https://github.com/midwan/amiberry/releases/download"
-AMIBERRY_VERSION="v5.1"
+AMIBERRY_VERSION="v5.2"
 AMIBERRY_BINARY="amiberry-$AMIBERRY_VERSION-rpi4-sdl2-64bit-rpios.zip"
 AMIBERRY_URL="$GH_URL/$AMIBERRY_VERSION/$AMIBERRY_BINARY"
 
 header "Executing: 'setup_amiberry.sh'"
 
 ## Creating download folder if not exists.
-if [[ ! -d $EMIGA_DIR/downloads ]]; then 
+if [[ ! -d $EMIGA_DIR/downloads ]]; then
    printf "~> Creating folder: '$EMIGA_DIR/downloads/'... "
    mkdir -p $EMIGA_DIR/downloads > /dev/null 2>&1
    status
 fi
 
 ## Create public/configs folder if not exists.
-if [[ ! -d $EMIGA_DIR/public/configs ]]; then 
+if [[ ! -d $EMIGA_DIR/public/configs ]]; then
    printf "~> Creating folder: '$EMIGA_DIR/public/configs/'... "
    mkdir -p $EMIGA_DIR/public/configs > /dev/null 2>&1
    status
 fi
 
 ## Create public/kickstarts folder if not exists.
-if [[ ! -d $EMIGA_DIR/public/kickstarts ]]; then 
+if [[ ! -d $EMIGA_DIR/public/kickstarts ]]; then
    printf "~> Creating folder: '$EMIGA_DIR/public/kickstarts/'... "
    mkdir -p $EMIGA_DIR/public/kickstarts > /dev/null 2>&1
    status
@@ -51,12 +51,14 @@ if [[ ! -d $EMIGA_DIR/amiberry ]]; then
    unzip -q "$EMIGA_DIR/downloads/$AMIBERRY_BINARY" -d "$EMIGA_DIR/amiberry" > /dev/null 2>&1
    status
 
-   ## Adapt Amiberry folder strucure.
-   printf "~> Adapt Amiberry folder structure... "
-   mv $EMIGA_DIR/amiberry/*/* $EMIGA_DIR/amiberry/ > /dev/null 2>&1
-   rm -r $EMIGA_DIR/amiberry/amiberry*/ > /dev/null 2>&1
-   status
-   
+   ## Check if Amiberry files and folders are contained in a subfolder.
+   if [[ ! -f $EMIGA_DIR/amiberry/amiberry || ! $EMIGA_DIR/amiberry/conf ]]; then
+      printf "~> Adapt Amiberry folder structure... "
+      mv $EMIGA_DIR/amiberry/*/* $EMIGA_DIR/amiberry/ > /dev/null 2>&1
+      rm -r $EMIGA_DIR/amiberry/amiberry*/ > /dev/null 2>&1
+      status
+   fi
+
    ## Copy Amiberry AROS kickstart roms into public/kickstarts folder.
    printf "~> Copy Amiberry AROS kickstart roms to '$EMIGA_DIR/public/kickstarts/'... "
    cp $EMIGA_DIR/amiberry/kickstarts/aros* $EMIGA_DIR/public/kickstarts > /dev/null 2>&1
@@ -64,7 +66,7 @@ if [[ ! -d $EMIGA_DIR/amiberry ]]; then
 fi
 
 ## Create symlink for Amiberry conf/ folder.
-if [[ -d $EMIGA_DIR/amiberry/conf ]]; then 
+if [[ -d $EMIGA_DIR/amiberry/conf ]]; then
    printf "~> Create symlink: 'amiberry/conf/' -> '$EMIGA_DIR/public/configs/'... "
    mv $EMIGA_DIR/amiberry/conf $EMIGA_DIR/amiberry/conf.org > /dev/null 2>&1
    ln -fs $EMIGA_DIR/public/configs $EMIGA_DIR/amiberry/conf > /dev/null 2>&1
@@ -72,7 +74,7 @@ if [[ -d $EMIGA_DIR/amiberry/conf ]]; then
 fi
 
 ## Create symlink for Amiberry kickstarts/ folder.
-if [[ -d $EMIGA_DIR/amiberry/kickstarts ]]; then 
+if [[ -d $EMIGA_DIR/amiberry/kickstarts ]]; then
    printf "~> Create symlink: 'amiberry/kickstarts/' -> '$EMIGA_DIR/public/kickstarts/'... "
    mv $EMIGA_DIR/amiberry/kickstarts $EMIGA_DIR/amiberry/kickstarts.org > /dev/null 2>&1
    ln -fs $EMIGA_DIR/public/kickstarts $EMIGA_DIR/amiberry/kickstarts > /dev/null 2>&1
